stop shitting in my console
The camera rigs don't get destroyed during normal play, so now my console is safe from being shit in while changing avatar.

This likely will fuck with how the distance scaling works, but it *should* be a good change. The camera rigs are more likely to be closer to the nameplates than the root of the avatar to the nameplate anyways, so it makes more sense to scale by.

This does mean though that crouching or being small next to a tall player will likely show the nameplate. It would be nice if the distance took into account height of the remote player, but that would be a funni fix.